Assignment6/pyproject.toml
[tool.poetry]
name = "rotatingonlyopen64"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
simonspeckciphers = "^1.0"
pycryptodome = "^3.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
